Add PromptProcessor to manage prompt processing more effectively
This change introduces a new `PromptProcessor` class to handle the prompt processing logic. It encapsulates functionality such as processing shell commands and ERB, making the code cleaner and more modular. This also allows for easier future modifications and enhancements to prompt handling.
