+++ lib/aia/prompt_handler.rb
@@ -3,6 +3,7 @@
 # This file handles prompt management for the AIA application.
 
 require 'prompt_manager'
+require_relative 'prompt_processor'
 require 'prompt_manager/storage/file_system_adapter'
 require 'erb'
 
@@ -28,6 +29,7 @@ module AIA
           config.prompt_extension = '.txt'  # default
           config.params_extension = '.json' # default
         end.new
+      @prompt_processor = PromptProcessor.new(config)
     end
 
     # Retrieves and processes a prompt by its ID, optionally prepending a role.
@@ -56,125 +58,7 @@ module AIA
     # @param prompt [PromptManager::Prompt, String] the prompt to process
     # @return [String] the processed prompt text
     def process_prompt(prompt)
-      # Deep copy the prompt to avoid modifying the original
-      if prompt.is_a?(PromptManager::Prompt)
-        # Process shell commands if enabled
-        if @config.shell
-          prompt.text = prompt.text.gsub(/$((.*?))/) { `#{Regexp.last_match(1)}`.chomp }
-        end
-
-        # Process ERB if enabled
-        if @config.erb
-          prompt.text = ERB.new(prompt.text).result(binding)
-        end
-
-        # Build the prompt with parameters
-        text = prompt.to_s
-
-        # Process directives after building the prompt
-        directives = prompt.directives
-        text = process_collected_directives(text, directives)
-
-        # Add terse instruction if requested
-        if @config.terse
-          text += "
-
-Please be terse in your response."
-        end
-
-        text
-      else
-        # Just a plain text prompt
-        text = prompt.dup
-
-        # Process shell commands if enabled
-        if @config.shell
-          text = text.gsub(/$((.*?))/) { `#{Regexp.last_match(1)}`.chomp }
-        end
-
-        # Process ERB if enabled
-        if @config.erb
-          text = ERB.new(text).result(binding)
-        end
-
-        # Add terse instruction if requested
-        if @config.terse
-          text += "
-
-Please be terse in your response."
-        end
-
-        text
-      end
-    end
-
-    private
-
-    # Processes collected directives within the prompt text, executing
-    # commands or including files as specified.
-    #
-    # @param text [String] the prompt text containing directives
-    # @param directives [Hash] the directives to process
-    # @return [String] the processed text with directives applied
-    def process_collected_directives(text, directives)
-      directives.each do |directive, args|
-        case directive
-        when "config"
-          # Process config directive
-          key, value = args.split(/\s*=\s*/, 2)
-          @config[key.strip.to_sym] = parse_value(value.strip)
-        when "include"
-          # Include another file
-          file_path = args.strip
-          if File.exist?(file_path)
-            # Replace the directive line with file contents
-            text = text.gsub(%r{//include #{Regexp.escape(args)}}, File.read(file_path))
-          else
-            # Replace with error message
-            text = text.gsub(%r{//include #{Regexp.escape(args)}}, "# Error: File not found: #{file_path}")
-          end
-        when "shell"
-          # Execute shell command
-          cmd_output = `#{args}`.chomp
-          # Replace the directive line with command output
-          text = text.gsub(%r{//shell #{Regexp.escape(args)}}, cmd_output)
-        when "ruby"
-          # Execute Ruby code
-          result = eval(args)
-          # Replace the directive line with result
-          text = text.gsub(%r{//ruby #{Regexp.escape(args)}}, result.to_s)
-        when "next"
-          # Set next prompt
-          @config.next = args.strip
-        when "pipeline"
-          # Set pipeline
-          @config.pipeline = args.strip.split(',')
-        end
-      end
-
-      text
-    end
-
-    # Parses a value from a string, converting it to the appropriate type
-    # (e.g., boolean, integer, array).
-    #
-    # @param value [String] the value to parse
-    # @return [Object] the parsed value
-    def parse_value(value)
-      case value.downcase
-      when 'true'
-        true
-      when 'false'
-        false
-      when /^\d+$/
-        value.to_i
-      when /^\d+\.\d+$/
-        value.to_f
-      when /^\[.*\]$/
-        value[1..-2].split(',').map(&:strip)
-      else
-        value
-      end
+      @prompt_processor.process(prompt)
     end
   end
 end
